# rhythmiqcontrolplaneservice/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Profile API
  version: 1.0.0
paths:
  /profiles:
    post:
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProfileResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationException'
        '409':
          description: Conflict error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
    get:
      summary: List all profiles
      responses:
        '200':
          description: Profiles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProfilesResponse'
  /profiles/{profileId}:
    get:
      summary: Get a profile by ID
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfileResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundException'
    put:
      summary: Update a profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationException'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundException'
components:
  schemas:
    CreateProfileRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        email:
          type: string
    CreateProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    GetProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    UpdateProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    ListProfilesResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            $ref: '#/components/schemas/GetProfileResponse'
    ValidationException:
      type: object
      properties:
        message:
          type: string
    ConflictException:
      type: object
      properties:
        message:
          type: string
    NotFoundException:
      type: object
      properties:
        message:
          type: string